services/fire-risk-service/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Caching and Session Management
redis==5.0.1
hiredis==2.2.3

# Machine Learning and Data Science
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
numpy==1.25.2
pandas==2.1.4
scipy==1.11.4

# Geospatial Processing
geopandas==0.14.1
shapely==2.0.2
rasterio==1.3.9
fiona==1.9.5

# Time Series and Weather Data
xarray==2023.12.0
netcdf4==1.6.5
h5netcdf==1.3.0

# Deep Learning and Google Research Integration
torch==2.1.2
torchvision==0.16.2
tensorflow==2.15.0
google-cloud-aiplatform==1.38.1
apache-beam[gcp]==2.52.0

# API and HTTP Clients
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Message Queue and Streaming
kafka-python==2.0.2
aiokafka==0.8.11

# Monitoring and Metrics
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Configuration and Environment
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3
structlog==23.2.0
rich==13.7.0

# Model Serving and MLOps
mlflow==2.8.1
joblib==1.3.2
cloudpickle==3.0.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1

# Data Validation
cerberus==1.3.5
marshmallow==3.20.1